+++ src/libc/include/owi.h
@@ -41,6 +41,12 @@ owi_bool(void);
 #ifdef __OWI_INTERNALS
 __attribute__((import_module("symbolic"))) __attribute__((import_name("invisible_bool_symbol"))) int
 owi_invisible_bool();
+
+__attribute__((import_module("symbolic"))) __attribute__((import_name("open_scope"))) void
+owi_open_scope(const char *name);
+
+__attribute__((import_module("symbolic"))) __attribute__((import_name("end_scope"))) void
+owi_end_scope();
 #endif
 
 __attribute__((import_module("symbolic"), import_name("assume"))) void