+++ src/libc/src/klee.c
@@ -1,10 +1,13 @@
 #include <assert.h>
 #include <klee.h>
-#include <owi.h>
 #include <stddef.h>
 #include <stdint.h>
 #include <stdlib.h>
 #include <string.h>
+
+#define __OWI_INTERNALS
+#include <owi.h>
+#undef __OWI_INTERNALS
 
 unsigned char *make_in(size_t size) {
   unsigned char *t = malloc(size);
@@ -32,13 +35,23 @@ void klee_define_fixed_object(void *addr, size_t nbytes) {
 }
 
 void klee_make_symbolic(void *addr, size_t nbytes, const char *name) {
+  owi_open_scope(name);
   memcpy(addr, make_in(nbytes), nbytes);
+  owi_end_scope();
 }
 
 int klee_range(int begin, int end, const char *name) {
-  return owi_range(begin, end);
+  owi_open_scope(name);
+  int range = owi_range(begin, end);
+  owi_end_scope();
+  return range;
 }
-int klee_int(const char *name) { return owi_i32(); }
+int klee_int(const char *name) {
+  owi_open_scope(name);
+  int i = owi_i32();
+  owi_end_scope();
+  return i;
+}
 
 void klee_silent_exit(int status) { owi_exit(status); }
 void klee_abort() { owi_abort(); }