+++ test/c/scopes.t
@@ -1,0 +1,9 @@
+  $ owi c scopes.c -O0 --no-value
+  owi: [ERROR] Assert failure: false
+  model {
+    scope "scope 1" {
+      symbol symbol_0 i32
+    }
+  }
+  owi: [ERROR] Reached problem!
+  [13]